Move IdentityFilter to separate module, fix for new lints
In preparation for adding some new features, this change
simplifies authenticate() function by passing in an &IdentityFilter.

Also, rustc has gotten some new ideas on declaring lifetimes
explicitly, adding that to shut up the warnings.

As I moved some tests around, I realised that I wanted to
have a central place for test stuff
