+++ examples/authenticator.rs
@@ -1,6 +1,7 @@
 use anyhow::Result;
 use log::info;
 use pam_ssh_agent::authenticate;
+use pam_ssh_agent::filter::IdentityFilter;
 use ssh_agent_client_rs::Client;
 use std::env;
 use std::path::Path;
@@ -14,7 +15,8 @@ fn main() -> Result<()> {
     env_logger::builder()
         .filter_level(log::LevelFilter::Info)
         .init();
-    let result = authenticate(authorized_keys_path.as_str(), None, client, "")?;
+    let filter = IdentityFilter::from_files(Path::new(authorized_keys_path.as_str()), None)?;
+    let result = authenticate(&filter, client, "")?;
     if result {
         info!("the ssh agent at {path} signed a random message as validated by {authorized_keys_path}");
     } else {
