+++ src/pam.rs
@@ -1,0 +1,2 @@
+use pam::module::PamHandle;
+
